pgkernelgateway/src/test/javascript/spec/app/entities/agent-type/agent-type-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { Observable, of } from 'rxjs';

import { PgkernelgatewayTestModule } from '../../../test.module';
import { AgentTypeUpdateComponent } from 'app/entities/agent-type/agent-type-update.component';
import { AgentTypeService } from 'app/entities/agent-type/agent-type.service';
import { AgentType } from 'app/shared/model/agent-type.model';

describe('Component Tests', () => {
  describe('AgentType Management Update Component', () => {
    let comp: AgentTypeUpdateComponent;
    let fixture: ComponentFixture<AgentTypeUpdateComponent>;
    let service: AgentTypeService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [PgkernelgatewayTestModule],
        declarations: [AgentTypeUpdateComponent],
        providers: [FormBuilder]
      })
        .overrideTemplate(AgentTypeUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(AgentTypeUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(AgentTypeService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new AgentType(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new AgentType();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
